\name{metaDF2Matrix.metabinDF}
\alias{metaDF2Matrix.metabinDF}

\title{
Convert 'metabinDF' Objects to Matrices
}

\description{  
A function that converts \code{'metabinDF'} objects into a character matrix
used for plotting.
}

\usage{
\method{metaDF2Matrix}{metabinDF}(df, order, newCols = NULL, roundCols = NULL,
                                  stats = list(hetero = makeStatsDesc(labelNames = c("I2", "tau2", "p"))), 
                                  groupStats = NULL, newLabel = NULL, 
                                  colNames, groupLab, hgap = NULL, vgap, drop, ...) 
}



\arguments{
  \item{df}{An object of class 'metabinDF'}
  
  \item{order}{A vector of column names in the order that it should
  appear in the matrix; By default, The columns 'study' and 'effect'
  will appear. The details of some extra default columns with special
  formats are given under 'Details'.}
  
  \item{newCols}{A list of new columns made from existing columns in the
  data frame. Can use function \code{\link{makeColDesc}} to combine and
  round columns.}
  
  \item{roundCols}{A named vector with values of rounding to the nearest
  decimal place of specified column names. The details of default
  rounding are given under 'Details'.}
  
  \item{stats}{A list of descriptions or strings for heterogeneity
  statistics. Can use the functions \code{\link{makeStatsDesc}} and
  \code{\link{makeLabelDesc}} to make descriptions that specify format
  and select statistics. More details are given under 'Details'.}
  
  \item{newLabel}{A list of new labels made from existing hetero in the
  data frame. Can use function \code{\link{makeLabelDesc}} to make label
  of hetero with specific format.}
  
  \item{groupStats}{A list of descriptions or string for heterogeneity
  statistics for overall summary in grouped studies. Can use function
  \code{\link{makeStatsDesc}} and \code{\link{makeLabelDesc}} to make
  descriptions that specify format and select statistics. By default, it
  is the same as the stats for individualgroups.}

  \item{colNames}{A vector of column names for the ordered matrix.}

  \item{groupLab}{A vector of labels for subgroups and overall effect.}

  \item{hgap}{A vector of row numbers to add horizontal space.}

  \item{vgap}{A vector of column numbers to add vertical spaces.}

  \item{drop}{A vector of either column numbers or rownames to drop rows
  from the matrix.}

  \item{...}{Not used.}
  
}
\details{  
There is one extra column, 'ci', provided for users to call in the
argument \code{order}. It is a column with special format, '[e.lower,
e.upper]', for the columns 'e.lower' and 'e.upper'.


Rounding can be applied using the argument \code{roundCols} for all
numeric columns except for the one with special format like 'ci' or or
the ones with cusomised format through the argument
\code{newCols}. These types of columns must be done through the argument
\code{newCols} using function \code{\link{makeColDesc}}. The default
roundings for columns are n.e = 0, event.e = 0, n.c = 0, event.c = 0,
effect = 2, se = 2, w.fixed = 0 or 1, w.random = 0 or 1, mean = 2, lower
= 2, upper = 2, e.lower = 2, e.upper = 2, and all other columns = 2.


The argument \code{stats} accept four classes of object, 'numeric',
'character', 'statsDesc' and 'labelDesc'. The later two classes of
object are generated by the functions \code{\link{makeStatsDesc}} and
\code{\link{makeLabelDesc}} respectively. For 'statsDesc', the
description for heterogeneity statistics are made from existing labels
setting in \code{\link{metaDF2Matrix.metabin}}. The default labels are
'Q', 'p', 'df', 'tau2', 'H', 'H.ci', 'I2', 'I2.ci', 'Q.CMH'. New labels
can be created by the argument \code{newLabel} through
\code{\link{makeLabelDesc}}. For 'labelDesc', the description for
heterogeneity statistics are made from existing hetero from the data
frame.
}

\value{

  An object of class '\code{metabinM}' and '\code{metaM}' with a list of following
  components:

  \item{matrix}{A matrix set out as to be drawn in the forest plot.}


  \item{plotDF}{A data frame with plotting information for dawing the
   confidence interval graph.}

}


\seealso{
\code{\link{meta2DF.metabin}}, \code{\link{drawMeta.metabinM}}
}

\examples{
## metabin: single group
library(meta)
data(Olkin95)
meta <- metabin(event.e, n.e, event.c, n.c, data=Olkin95, 
                subset=c(41,47,51,59), sm="RR", method="I")                                

Data <- meta2DF(meta, title = "Thrombolytic Therapy" , subtitle = "Olkin I (1995)",
                rowOrder = "effect", decreasing = TRUE)

matrix1 <- metaDF2Matrix(Data, order = c("study", "effect", "ci", "weight"),
                         newCols = list(weight = makeColDesc(format = "\%.0f\%\%", colNames = "w.fixed")), 
                         stats = list(makeStatsDesc(labelNames = c("I2", "p", "tau2")),
                                      makeStatsDesc(labelNames = c("Q", "df"), emptyHeading = TRUE),
                                      makeLabelDesc(format = "Test for Overall Effect: Q = \%.f",
                                                    heteroNames = "Q.CMH")))
                                                    
matrix1

## metabin: grouped studies
Olkin95.2 <- Olkin95[1:10, ]
meta <- metabin(event.e, n.e, event.c, n.c, data=Olkin95.2, studlab = author,
                sm = "RR", method = "I", byvar = c(1, 2, 1, 1, 2, 2, 2, 3, 3, 3),
                bylab = "group")

Data <- meta2DF(meta, title = "Thrombolytic Therapy", subtitle = "Group Studies for Olkin95", 
                rowOrder = "effect", decreasing = TRUE) 

matrix2 <- metaDF2Matrix(Data, groupLab = c("Group One", "Group Two", "Group Three", "Overall"),
                         groupStats = list(makeStatsDesc(labelNames = c("I2", "p", "tau2")),
                                           makeStatsDesc(labelNames = "Q.CMH", emptyHeading = TRUE)))
                                           
matrix2
}

\keyword{metaDF2Matrix}

