\name{metaDF2Matrix.metacontDF}
\alias{metaDF2Matrix.metacontDF}
\title{
Convert 'metacontDF' Objects to Matrices
}
\description{
A function that converts \code{'metacontDF'} objects into a character matrix 
used for plotting.
}
\usage{
\method{metaDF2Matrix}{metacontDF}(df, order, newCols = NULL, roundCols = NULL,
                                  stats = list(hetero = makeStatsDesc(labelNames = c("I2", "tau2", "p"))), 
                                  groupStats = NULL, newLabel = NULL, 
                                  colNames, groupLab, hgap = NULL, vgap, drop, ...) 
}
\arguments{
  \item{df}{An object of class 'metacontDF'}
  
  \item{order}{A vector of column names in the order that it should
  appear in the matrix; By default, The columns 'study' and 'effect'
  will appear. The details of some extra default columns with special
  formats are given under 'Details'.}
  
  \item{newCols}{A list of new columns made from existing columns in the
  data frame. Can use function \code{\link{makeColDesc}} to combine and
  round columns.}
  
  \item{roundCols}{A named vector with values of rounding to the nearest
  decimal place of specified column names. The details of default
  rounding are given under 'Details'.}
  
  \item{stats}{A list of descriptions or strings for heterogeneity
  statistics. Can use the functions \code{\link{makeStatsDesc}} and
  \code{\link{makeLabelDesc}} to make descriptions that specify format
  and select statistics. More details are given under 'Details'.}
  
  \item{newLabel}{A list of new labels made from existing hetero in the
  data frame. Can use function \code{\link{makeLabelDesc}} to make label
  of hetero with specific format.}
  
  \item{groupStats}{A list of descriptions or string for heterogeneity
  statistics for overall summary in grouped studies. Can use function
  \code{\link{makeStatsDesc}} and \code{\link{makeLabelDesc}} to make
  descriptions that specify format and select statistics. By default, it
  is the same as the stats for individualgroups.}

  \item{colNames}{A vector of column names for the ordered matrix.}

  \item{groupLab}{A vector of labels for subgroups and overall effect.}

  \item{hgap}{A vector of row numbers to add horizontal space.}

  \item{vgap}{A vector of column numbers to add vertical spaces.}

  \item{drop}{A vector of either column numbers or rownames to drop rows
  from the matrix.}

  \item{...}{Not used.}
}

\details{
There are three extra columns with special format, 'ci', 'msd.e' and
'msd.c' provided for users to call in the argument \code{order}.  The
default formats for the columns 'ci', 'msd.e' and 'msd.c' are '[lower,
upper]', 'mean.e(sd.e)' and 'mean.c(sd.c)' respectively.

Rounding can be applied using the argument \code{roundCols} for all
numeric columns except for the one with special format like 'ci',
'msd.e', 'msd.c' or or the ones with cusomised format through the
argument \code{newCols}. These types of columns must be done through the
argument \code{newCols} using function \code{\link{makeColDesc}}. The
default roundings for columns are n.e = 0, mean.e = 1, n.c = 0, mean.c =
1, sd.e = 2, sd.c = 2, effect = 2, w.fixed = 0 or 1, w.random = 0 or 1,
mean = 2, lower = 2, upper = 2, any other columns = 2.

The argument \code{stats} accept four classes of object, 'numeric',
'character', 'statsDesc' and 'labelDesc'. The later two classes of
object are generated by the functions \code{\link{makeStatsDesc}} and
\code{\link{makeLabelDesc}} respectively. For 'statsDesc', the
description for heterogeneity statistics are made from existing labels
setting in \code{\link{metaDF2Matrix.metabin}}. The default labels are
'Q', 'p', 'df', 'tau2', 'H', 'H.ci', 'I2', 'I2.ci'. New labels can be
created by the argument \code{newLabel} through
\code{\link{makeLabelDesc}}. For 'labelDesc', the description for
heterogeneity statistics are made from existing hetero from the data
frame.
}

\value{
An object of class '\code{metacontM}' and '\code{metaM}' with a list
of following components:
  \item{matrix}{A matrix set out as to be drawn in the forest plot.}
  \item{plotDF}{A data frame with plotting information for drawing the 
                confidence interval graph.}
}
\seealso{
\code{\link{meta2DF.metacont}},
\code{\link{drawMeta.metacontM}}
}
\examples{
library(meta)
data(Fleiss93cont)
## metacont: single group 
meta1 <- metacont(n.e, mean.e, sd.e, n.c, mean.c, sd.c, data=Fleiss93cont, 
                  sm="SMD")

Data1 <- meta2DF(meta1, title = "Mental Health Treatment",
                 rowOrder = "effect", decreasing = TRUE) 

matrix1 <- metaDF2Matrix(Data1, order = c("study", "msd.e", "msd.c", "effect", 
                                          "ci", "w.fixed", "w.random"),
                         roundCols = c("effect" = 3, "w.fixed" = 1, "w.random" = 1),
                         colNames = c("Study", "Experimental Group", "Control Group", "Effect",
                                      "95\% CI", "Weight(fixed)", "Weight(random)"))
                                      
matrix1

## metacont: grouped studies
meta2 <- metacont(n.e, mean.e, sd.e, n.c, mean.c, sd.c, data=Fleiss93cont, 
                  sm="SMD", byvar=c(1,2,1,1,2), bylab="group") 
                  
Data2 <- meta2DF(meta2, title = "Mental Health Treatment",
                 rowOrder = "effect", decreasing = TRUE) 
                       
matrix2 <- metaDF2Matrix(Data2, groupLab = c("Group One", "Group Two", "Overall"),
                         order = c("study", "msd.e", "msd.c", "effect",
                                   "ci", "w.fixed", "w.random"))

matrix2
}
\keyword{metaDF2Matrix}